Add behaviour for harvester, upgrader, and scout
This commit adds the basic behaviour for several roles in the Screeps AI.
The roles are written in the style of Object-oriented JS.

The build-queue module/class is responsible for deciding what gets built,
and it's state is stored in Memory. In the future, the queue will be managed
by another class.