Add Builder and basic room planning to AI
This commit adds a new role - builder, which is responsible for
building all the construction sites that have been layed out by the
room-planner. At the moment, the room-planner relies on a map of structures
and their position relative to the Spawn given to the planner. The plans are
divided based on the level of the controller (which places limitations on the
structures you can have.

Also fixed minor type comparison (!= becomes !==).